copy*(): use fs.stat with bigint option
