+++ lib/copy-sync/copy-sync.js
@@ -4,8 +4,7 @@ const fs = require('graceful-fs')
 const path = require('path')
 const mkdirpSync = require('../mkdirs').mkdirsSync
 const utimesSync = require('../util/utimes.js').utimesMillisSync
-
-const notExist = Symbol('notExist')
+const stat = require('../util/stat')
 
 function copySync (src, dest, opts) {
   if (typeof opts === 'function') {
@@ -22,12 +21,8 @@ function copySync (src, dest, opts) {
     see https://github.com/jprichardson/node-fs-extra/issues/269`)
   }
 
-  if (opts.checkPathsBeforeCopying === false) {
-    const { destStat } = checkStats(src, dest)
-    return handleFilterAndCopy(destStat, src, dest, opts)
-  }
-  const { srcStat, destStat } = checkPaths(src, dest)
-  checkParentPaths(src, srcStat, dest)
+  const { srcStat, destStat } = stat.checkPathsSync(src, dest)
+  stat.checkParentPathsSync(src, srcStat, dest)
   return handleFilterAndCopy(destStat, src, dest, opts)
 }
 
@@ -55,7 +50,7 @@ function getStats (destStat, src, dest, opts) {
 }
 
 function onFile (srcStat, destStat, src, dest, opts) {
-  if (destStat === notExist) return copyFile(srcStat, src, dest, opts)
+  if (!destStat) return copyFile(srcStat, src, dest, opts)
   return mayCopyFile(srcStat, src, dest, opts)
 }
 
@@ -101,7 +96,7 @@ function copyFileFallback (srcStat, src, dest, opts) {
 }
 
 function onDir (srcStat, destStat, src, dest, opts) {
-  if (destStat === notExist) return mkDirAndCopy(srcStat, src, dest, opts)
+  if (!destStat) return mkDirAndCopy(srcStat, src, dest, opts)
   if (destStat && !destStat.isDirectory()) {
     throw new Error(`Cannot overwrite non-directory '${dest}' with directory '${src}'.`)
   }
@@ -121,19 +116,17 @@ function copyDir (src, dest, opts) {
 function copyDirItem (item, src, dest, opts) {
   const srcItem = path.join(src, item)
   const destItem = path.join(dest, item)
-  const check = opts.checkPathsBeforeCopying === false ? checkStats : checkPaths
-  const { destStat } = check(srcItem, destItem)
+  const { destStat } = stat.checkPathsSync(srcItem, destItem)
   return startCopy(destStat, srcItem, destItem, opts)
 }
 
 function onLink (destStat, src, dest, opts) {
   let resolvedSrc = fs.readlinkSync(src)
-
   if (opts.dereference) {
     resolvedSrc = path.resolve(process.cwd(), resolvedSrc)
   }
 
-  if (destStat === notExist) {
+  if (!destStat) {
     return fs.symlinkSync(resolvedSrc, dest)
   } else {
     let resolvedDest
@@ -149,14 +142,14 @@ function onLink (destStat, src, dest, opts) {
     if (opts.dereference) {
       resolvedDest = path.resolve(process.cwd(), resolvedDest)
     }
-    if (isSrcSubdir(resolvedSrc, resolvedDest)) {
+    if (stat.isSrcSubdir(resolvedSrc, resolvedDest)) {
       throw new Error(`Cannot copy '${resolvedSrc}' to a subdirectory of itself, '${resolvedDest}'.`)
     }
 
     // prevent copy if src is a subdir of dest since unlinking
     // dest in this case would result in removing src contents
     // and therefore a broken symlink would be created.
-    if (fs.statSync(dest).isDirectory() && isSrcSubdir(resolvedDest, resolvedSrc)) {
+    if (fs.statSync(dest).isDirectory() && stat.isSrcSubdir(resolvedDest, resolvedSrc)) {
       throw new Error(`Cannot overwrite '${resolvedDest}' with '${resolvedSrc}'.`)
     }
     return copyLink(resolvedSrc, dest)
@@ -166,57 +159,6 @@ function onLink (destStat, src, dest, opts) {
 function copyLink (resolvedSrc, dest) {
   fs.unlinkSync(dest)
   return fs.symlinkSync(resolvedSrc, dest)
-}
-
-// return true if dest is a subdir of src, otherwise false.
-// It only checks the path strings.
-function isSrcSubdir (src, dest) {
-  const srcArr = path.resolve(src).split(path.sep).filter(i => i)
-  const destArr = path.resolve(dest).split(path.sep).filter(i => i)
-  return srcArr.reduce((acc, cur, i) => acc && destArr[i] === cur, true)
-}
-
-function checkStats (src, dest) {
-  const srcStat = fs.statSync(src)
-  let destStat
-  try {
-    destStat = fs.statSync(dest)
-  } catch (err) {
-    if (err.code === 'ENOENT') return { srcStat, destStat: notExist }
-    throw err
-  }
-  return { srcStat, destStat }
-}
-
-// recursively check if dest parent is a subdirectory of src.
-// It works for all file types including symlinks since it
-// checks the src and dest inodes. It starts from the deepest
-// parent and stops once it reaches the src parent or the root path.
-function checkParentPaths (src, srcStat, dest) {
-  const destParent = path.dirname(dest)
-  if (destParent && (destParent === path.dirname(src) || destParent === path.parse(destParent).root)) return
-  let destStat
-  try {
-    destStat = fs.statSync(destParent)
-  } catch (err) {
-    if (err.code === 'ENOENT') return
-    throw err
-  }
-  if (destStat.ino && destStat.ino === srcStat.ino) {
-    throw new Error(`Cannot copy '${src}' to a subdirectory of itself, '${dest}'.`)
-  }
-  return checkParentPaths(src, srcStat, destParent)
-}
-
-function checkPaths (src, dest) {
-  const { srcStat, destStat } = checkStats(src, dest)
-  if (destStat.ino && destStat.ino === srcStat.ino) {
-    throw new Error('Source and destination must not be the same.')
-  }
-  if (srcStat.isDirectory() && isSrcSubdir(src, dest)) {
-    throw new Error(`Cannot copy '${src}' to a subdirectory of itself, '${dest}'.`)
-  }
-  return { srcStat, destStat }
 }
 
 module.exports = copySync
